rename `remot` package to `remote` in exercise repository
