+++ composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/di/remoteModule.kt
@@ -4,8 +4,8 @@ import com.cairosquad.evolvefit.repository.authentication.remote.AuthenticationR
 import com.cairosquad.evolvefit.repository.authentication.remote.AuthenticationRemoteDataSourceImpl
 import com.cairosquad.evolvefit.repository.equipment.remot.EquipmentRemoteDataSourceImpl
 import com.cairosquad.evolvefit.repository.equipment.remot.EquipmentsRemoteDataSource
-import com.cairosquad.evolvefit.repository.exercise.remot.ExerciseRemoteDataSource
-import com.cairosquad.evolvefit.repository.exercise.remot.ExerciseRemoteDataSourceImpl
+import com.cairosquad.evolvefit.repository.exercise.remote.ExerciseRemoteDataSource
+import com.cairosquad.evolvefit.repository.exercise.remote.ExerciseRemoteDataSourceImpl
 import com.cairosquad.evolvefit.repository.utils.RefreshTokenProvider
 import com.cairosquad.evolvefit.repository.utils.provideHttpClient
 import com.cairosquad.evolvefit.repository.workout.remote.WorkoutRemoteDataSource
