--- composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/repository/exercise/remot/ExerciseRemoteDataSource.kt
@@ -1,8 +1,0 @@
-package com.cairosquad.evolvefit.repository.exercise.remot
-
-import com.cairosquad.evolvefit.repository.exercise.remot.dto.ExerciseDto
-import com.cairosquad.evolvefit.repository.exercise.remot.dto.ExerciseResponseDto
-
-interface ExerciseRemoteDataSource {
-    suspend fun createExercise(exercise: ExerciseDto): ExerciseResponseDto
-}
