+++ composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/repository/exercise/remote/ExerciseRemoteDataSource.kt
@@ -1,0 +1,8 @@
+package com.cairosquad.evolvefit.repository.exercise.remote
+
+import com.cairosquad.evolvefit.repository.exercise.remote.dto.ExerciseDto
+import com.cairosquad.evolvefit.repository.exercise.remote.dto.ExerciseResponseDto
+
+interface ExerciseRemoteDataSource {
+    suspend fun createExercise(exercise: ExerciseDto): ExerciseResponseDto
+}
